ODE-41 Improving the reliability of the Importer class
Tested changes made by @c-ianculescu (PR #21) in a docker container. Resolved some issues with directory paths and files handling. Encountered with an issue that was caused by the new file not being fully written/closed before we attempt to process it. Added a retry loop to make sure that the file is processed when all the data is there. This is a workaround until we find a better way to handle this situation.