Extract Kafka properties from OdeProperties as a first step to breaking apart the monstrosity  (#110)
* Migrate kafka configuration values to ODEKafkaProperties
This is the first part of the work needed to separate the OdeProperties into multpile more manageable chunks. It introduces the usage of application.yaml and @Autowired in tests to more fully utilize the Spring framework.

* remove @Component annotation and accept linter errors
Adding the @Component annotation to the *Receivers caused double instantiations and port binding issues

* migrate from application.properties to application.yaml and introducde host-ip config val

* add make restart target for easier, faster restarts that don't require a full image rebuild

* add missing getHostId test for odekafkaproperties

* move odekafkaproperties to kafka package

* correctly declare restart target as PHONY in Makefile

* move OdeKafkaPropertiesTest to correct test package

* add tests to OdePropertiesTest for hostIP and remove unused imports and variables

* moved hostID from OdeKafkaProperties to AppContext to keep scope of new class small and directed

* correct application.yaml's disabled-topics block

* correct formatting in Asn1*JSONTest.java files and fix Sonarlint errors
All swaps of argument positions in the mentioned tests are to align the function calls to the defition: assert*(expected, actual). Before this swap, the call was made in the tests like assert*(actual, expected) which has the same outcome when tests *pass* but is confusing when tests do not pass because the failure output will have the actual and expected values swapped in the message.

* add sprint.http.multipart config values to application.yaml

* Add TODO for @mcook42 to update README.md with corrected instructions once OdeProperties refactor is completed