fix: set elevation type to number in schema-map.json
The DE_Elevation data element represents the geographic position above or below the reference ellipsoid (typically WGS-84). The number has a resolution of 1 decimeter and represents an asymmetric range of positive and negative values. Any elevation higher than +6143.9 m is represented as +61439. Any elevation lower than -409.5 m is represented as -4095. If the sending device does not know its elevation, it shall encode the Elevation data element with -4096.