Migrate ToJsonController to Spring Kafka (#155)
* test: add canConvertPojoToJSON to ToJsonServiceControllerTest

* refactor: rename ToJsonServiceController to BSMPojoToJSONListener

Renamed the class and related test file to reflect its responsibility more accurately. Updated all references to align with the new class name. This improves clarity and consistency in the codebase.

* refactor: BSMPojoToJSONListener consumption now using Spring Kafka

The tests fail because I have not yet configured the correct KafkaTemplate for message production. This just proves we can consume the data without error

* refactor: BSMPojoToJSONListener production now using Spring Kafka

This uses a blend of Spring Kafka and our previous JsonUtils setup. The next step is to migrate to using the DI provided ObjectMapper to convert to the JSON representation during message production

* refactor: BSMPojoToJSONListener production using DI-provided ObjectMapper

* test: correctly close test consumer in BSMPojoToJSONListenerTest

* chore: remove deprecated and unused classes

Deleted `OdeBsmDeserializer`, `OdeBsmSerializer`, `OdeDataPublisher`, and related test files, as they were deprecated or no longer in use. Also marked `MessagingDeserializer` as `final` for better immutability and clarity.

* test: properly close out test consumers in Spring Kafka tests

* chore: delete unused ToJsonConverter.java

* chore: delete unused AbstractSubPubTransformer.java

* docs: add detailed JavaDocs to BSMPojoToJSONListener class

Added comprehensive JavaDocs to the BSMPojoToJSONListener class, including descriptions for the class itself, its constructor, and the Kafka listener method. These improvements clarify the purpose, functionality, and expected behavior of the class and its components.

* refactor: serialization to directly use Kryo API

Removed `SerializationUtils` and its associated test class, replacing them with direct usage of the Kryo library in `MessagingSerializer` and `MessagingDeserializer`. Updated serialization and deserialization logic to include logging and streamlined test cases for improved clarity and functionality. Separating Serialization and Deserialization responsibilities in to their respective classes better indicates the intentions of the classes.

* docs: correct references in javadocs

* chore: remove dead comment from BSMPojoToJSONListenerTest

* perf: update serializer to use Output.toBytes for accurate byte array

Replaced `output.getBuffer()` with `output.toBytes` to ensure accurate serialization and avoid including extra byte padding produced to the topic. this will save significant space on the topic as BSM data is one of the most frequent data types produced/consumed by the ODE.